# Conflicts:
#	routes/web.php

+++ routes/web.php
@@ -36,29 +36,33 @@ Route::controller(AdminController::class)->group(function () {
     Route::post('/admin/deconnexion', 'logout')->name('admin.logout')->middleware(RedirectIfNotAdmin::class);
     Route::get('/admin/dashboard', 'index')->name('admin.dashboard')->middleware(RedirectIfNotAdmin::class);
 
-    //CLIENTS
-    Route::get('/admin/clients', 'clients')->name('admin.clients')->middleware(RedirectIfNotAdmin::class);
-    Route::post('/admin/clients/add', 'addClient')->name('admin.clients.add')->middleware(RedirectIfNotAdmin::class);
-    Route::get('/admin/clients/autocomplete', 'autocomplete')->name('admin.clients.autocomplete')->middleware(RedirectIfNotAdmin::class);
-    Route::get('/admin/transactions', 'transactions')->name('admin.transactions')->middleware(RedirectIfNotAdmin::class);
+//CLIENTS
+Route::get('/admin/clients', [AdminController::class, 'clients'])->name('admin.clients')->middleware(RedirectIfNotAdmin::class);
+Route::post('/admin/clients/activate/{id}', [AdminController::class, 'activate'])->name('admin.clients.activate')->middleware(RedirectIfNotAdmin::class);
+Route::post('/admin/clients/deactivate/{id}', [AdminController::class, 'deactivate'])->name('admin.clients.deactivate')->middleware(RedirectIfNotAdmin::class);
+Route::get('/admin/clients/{id}', [AdminController::class, 'show'])->name('admin.clients.show')->middleware(RedirectIfNotAdmin::class);
+Route::put('/admin/clients/{id}', [AdminController::class, 'update'])->name('admin.clients.update')->middleware(RedirectIfNotAdmin::class);
+Route::get('/admin/clients/autocomplete', [AdminController::class, 'autocomplete'])->name('admin.clients.autocomplete')->middleware(RedirectIfNotAdmin::class);
+Route::post('/admin/clients/add', [AdminController::class, 'addClient'])->name('admin.clients.add')->middleware(RedirectIfNotAdmin::class);
+Route::get('/admin/transactions', [AdminController::class, 'transactions'])->name('admin.transactions')->middleware(RedirectIfNotAdmin::class);
 
-    //FESTIVALS
-    Route::get('/admin/festivals', 'festivals')->name('admin.festivals')->middleware(RedirectIfNotAdmin::class);
-    Route::post('/admin/festivals/add', 'addFestival')->name('admin.festivals.add')->middleware(RedirectIfNotAdmin::class);
-    Route::delete('/admin/festivals/{id}', 'deleteFestival')->name('admin.festivals.delete')->middleware(RedirectIfNotAdmin::class);
-    Route::put('/admin/festivals/{id}', 'updateFestival')->name('admin.festivals.update')->middleware(RedirectIfNotAdmin::class);
+//FESTIVALS
+Route::get('/admin/festivals', [AdminController::class, 'festivals'])->name('admin.festivals')->middleware(RedirectIfNotAdmin::class);
+Route::post('/admin/festivals/add', [AdminController::class, 'addFestival'])->name('admin.festivals.add')->middleware(RedirectIfNotAdmin::class);
+Route::delete('/admin/festivals/{id}', [AdminController::class, 'deleteFestival'])->name('admin.festivals.delete')->middleware(RedirectIfNotAdmin::class);
+Route::get('/admin/festivals/{id}', [AdminController::class, 'showFestival'])->name('admin.festivals.show')->middleware(RedirectIfNotAdmin::class);
+Route::put('/admin/festivals/{id}', [AdminController::class, 'updateFestival'])->name('admin.festivals.update')->middleware(RedirectIfNotAdmin::class);
 
-    //PROMOTIONS
-    Route::get('/admin/promotions', 'promotions')->name('admin.promotions')->middleware(RedirectIfNotAdmin::class);
-    Route::get('/admin/promotions', 'getOffers')->name('admin.promotions')->middleware(RedirectIfNotAdmin::class);
-    Route::post('/admin/offers/add', 'addOffer')->name('admin.offers.add')->middleware(RedirectIfNotAdmin::class);
+//PROMOTIONS
+Route::get('/admin/promotions', [AdminController::class, 'promotions'])->name('admin.promotions')->middleware(RedirectIfNotAdmin::class);
+Route::get('/admin/promotions', [AdminController::class, 'getOffers'])->name('admin.promotions')->middleware(RedirectIfNotAdmin::class);
+Route::post('/admin/offers/add', [AdminController::class, 'addOffer'])->name('admin.offers.add')->middleware(RedirectIfNotAdmin::class);
 
-    //NOTIFICATIONS
-    Route::get('/admin/notifications', 'notifications')->name('admin.notifications')->middleware(RedirectIfNotAdmin::class);
+//NOTIFICATIONS
+Route::get('/admin/notifications', [AdminController::class, 'notifications'])->name('admin.notifications')->middleware(RedirectIfNotAdmin::class);
 
-    //ACTUALITES
-    Route::get('/admin/actualites', 'actualites')->name('admin.actualites')->middleware(RedirectIfNotAdmin::class);
-});
+//ACTUALITES
+Route::get('/admin/actualites', [AdminController::class, 'actualites'])->name('admin.actualites')->middleware(RedirectIfNotAdmin::class);
 
 //PASSWORD RESET
 Route::post('/password/email', [AuthController::class, 'sendResetLinkEmail'])->name('password.email');
